+++ packages/jest-transform/src/__tests__/script_transformer.test.js
@@ -407,6 +407,7 @@ describe('ScriptTransformer', () => {
     expect(writeFileAtomic.sync).toBeCalledTimes(2);
     expect(writeFileAtomic.sync).toBeCalledWith(result.sourceMapPath, mapStr, {
       encoding: 'utf8',
+      fsync: false,
     });
   });
 
@@ -438,7 +439,7 @@ describe('ScriptTransformer', () => {
     expect(writeFileAtomic.sync).toBeCalledWith(
       result.sourceMapPath,
       sourceMap,
-      {encoding: 'utf8'},
+      {encoding: 'utf8', fsync: false},
     );
   });
 
@@ -507,6 +508,7 @@ describe('ScriptTransformer', () => {
       JSON.stringify(map),
       {
         encoding: 'utf8',
+        fsync: false,
       },
     );
   });
@@ -574,6 +576,7 @@ describe('ScriptTransformer', () => {
       JSON.stringify(instrumentedCodeMap),
       {
         encoding: 'utf8',
+        fsync: false,
       },
     );
 
@@ -614,6 +617,7 @@ describe('ScriptTransformer', () => {
       JSON.stringify(instrumentedCodeMap),
       {
         encoding: 'utf8',
+        fsync: false,
       },
     );